+++ abi_py/tradingapis/bitbank_api/private_api.py
